requirements.txt
fastapi
uvicorn
python-docx
python-dotenv
PyMuPDF
httpx
diskcache
